Refactoring: Move readFile() to fs.cc
